added the code in our backend files notif.odel and notif.service and removed the extra file
